+++ src/colorer/parsers/ParserFactoryImpl.h
@@ -20,7 +20,7 @@ class ParserFactory::Impl
   Impl& operator=(Impl&&) = delete;
 
   void loadCatalog(const UnicodeString* catalog_path);
-  void loadHrcPath(const UnicodeString& location, const UnicodeString* base_path) const;
+  void loadHrcPath(const UnicodeString* location, const UnicodeString* base_path) const;
   void loadHrcSettings(const UnicodeString* location) const;
 
   [[nodiscard]]